--- app/controllers/api/homes_controller.rb
@@ -1,3 +1,0 @@
-class Api::HomesController < JSONAPI::ResourceController
-  include Pundit::ResourceController
-end
