--- app/controllers/api/rooms_controller.rb
@@ -1,3 +1,0 @@
-class Api::RoomsController < JSONAPI::ResourceController
-  include Pundit::ResourceController
-end
