+++ app/controllers/api/v1/homes_controller.rb
@@ -1,0 +1,3 @@
+class Api::V1::HomesController < JSONAPI::ResourceController
+  include Pundit::ResourceController
+end
