+++ app/controllers/api/v1/rooms_controller.rb
@@ -1,0 +1,3 @@
+class Api::V1::RoomsController < JSONAPI::ResourceController
+  include Pundit::ResourceController
+end
