--- app/resources/api/application_resource.rb
@@ -1,6 +1,0 @@
-module Api
-  class ApplicationResource < JSONAPI::Resource
-    include Pundit::Resource
-    immutable
-  end
-end
