--- app/resources/api/home_resource.rb
@@ -1,6 +1,0 @@
-module Api
-  class HomeResource < ApplicationResource
-    model_name 'Home'
-    attribute :name
-  end
-end
