+++ app/resources/api/v1/application_resource.rb
@@ -1,0 +1,8 @@
+module Api
+  module V1
+    class ApplicationResource < JSONAPI::Resource
+      include Pundit::Resource
+      immutable
+    end
+  end
+end
