+++ app/resources/api/v1/home_resource.rb
@@ -1,0 +1,8 @@
+module Api
+  module V1
+    class HomeResource < ApplicationResource
+      model_name 'Home'
+      attribute :name
+    end
+  end
+end
